requirement.txt
Package                        Version
------------------------------ -----------------------
absl-py                        1.1.0
adversarial-robustness-toolbox 1.8.1
decorator                      5.1.1
eagerpy                        0.30.0
flatbuffers                    1.12
fonttools                      4.33.3
foolbox                        2.3.0
gast                           0.2.2
gitdb                          4.0.9
GitPython                      3.1.27
google-auth                    2.8.0
google-auth-oauthlib           0.4.6
google-pasta                   0.2.0
grpcio                         1.46.3
h5py                           2.10.0
idna                           3.3
imageio                        2.19.5
importlib-metadata             4.11.4
Markdown                       3.3.7
MarkupSafe                     2.1.1
matplotlib                     3.5.2
matplotlib-inline              0.1.3
mysql-connector-python         8.0.33
numba                          0.56.0
numpy                          1.21.6
nvidia-pyindex                 1.0.9
oauthlib                       3.2.0
opencv-python                  4.6.0.66
opt-einsum                     3.3.0
packaging                      21.3
pandas                         1.3.5
parso                          0.8.3
pexpect                        4.8.0
pickleshare                    0.7.5
Pillow                         9.1.0
pip                            22.0.4
prompt-toolkit                 3.0.29
protobuf                       3.19.4
psutil                         5.9.3
ptyprocess                     0.7.0
pyasn1                         0.4.8
pyasn1-modules                 0.2.8
pyDeprecate                    0.3.2
Pygments                       2.12.0
pyparsing                      3.0.8
python-dateutil                2.8.2
pytz                           2022.1
PyWavelets                     1.1.1
PyYAML                         6.0
requests                       2.27.1
requests-oauthlib              1.3.1
rsa                            4.8
scikit-learn                   1.0.2
scipy                          1.7.3
setuptools                     62.1.0
six                            1.16.0
smmap                          5.0.0
torch                          1.8.0+cu111
torch-cluster                  1.5.9
torch-geometric                2.0.4
torch-scatter                  2.0.8
torch-sparse                   0.6.9
torch-spline-conv              1.2.1
torchaudio                     0.8.0
torchmetrics                   0.8.2
torchsummary                   1.5.1
torchvision                    0.9.0+cu111
tqdm                           4.64.0
traitlets                      5.1.1
typing_extensions              4.2.0
urllib3                        1.26.9
wcwidth                        0.2.5
Werkzeug                       2.1.2
wheel                          0.37.1
wrapt                          1.14.1
zipp                           3.8.0
